fix(fine-tuning): unknow references and formatting
